investment-tracker/requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
yfinance==0.2.31
pandas==2.0.3
numpy>=1.25.0
scikit-learn==1.3.0
matplotlib==3.7.2
plotly==5.16.1
dash==2.13.0
Werkzeug==2.2.3
email-validator==2.0.0.post2
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
WTForms==2.3.3


